// frontend/src/types/iconTypes.ts
export type AntDesignNames =
  | "stepforward"
  | "stepbackward"
  | "forward"
  | "banckward"
  | "caretright"
  | "caretleft"
  | "caretdown"
  | "caretup"
  | "rightcircle"
  | "leftcircle"
  | "upcircle"
  | "downcircle"
  | "rightcircleo"
  | "leftcircleo"
  | "upcircleo"
  | "downcircleo"
  | "verticleleft"
  | "verticleright"
  | "back"
  | "retweet"
  | "shrink"
  | "arrowsalt"
  | "doubleright"
  | "doubleleft"
  | "arrowdown"
  | "arrowup"
  | "arrowright"
  | "arrowleft"
  | "down"
  | "up"
  | "right"
  | "left"
  | "minussquareo"
  | "minuscircle"
  | "minuscircleo"
  | "minus"
  | "pluscircleo"
  | "pluscircle"
  | "plus"
  | "infocirlce"
  | "infocirlceo"
  | "info"
  | "exclamation"
  | "exclamationcircle"
  | "exclamationcircleo"
  | "closecircle"
  | "closecircleo"
  | "checkcircle"
  | "checkcircleo"
  | "check"
  | "close"
  | "customerservice"
  | "creditcard"
  | "codesquareo"
  | "book"
  | "barschart"
  | "bars"
  | "question"
  | "questioncircle"
  | "questioncircleo"
  | "pause"
  | "pausecircle"
  | "pausecircleo"
  | "clockcircle"
  | "clockcircleo"
  | "swap"
  | "swapleft"
  | "swapright"
  | "plussquareo"
  | "frown"
  | "menufold"
  | "mail"
  | "link"
  | "areachart"
  | "linechart"
  | "home"
  | "laptop"
  | "star"
  | "staro"
  | "filter"
  | "meho"
  | "meh"
  | "shoppingcart"
  | "save"
  | "user"
  | "videocamera"
  | "totop"
  | "team"
  | "sharealt"
  | "setting"
  | "picture"
  | "phone"
  | "paperclip"
  | "notification"
  | "menuunfold"
  | "inbox"
  | "lock"
  | "qrcode"
  | "tags"
  | "tagso"
  | "cloudo"
  | "cloud"
  | "cloudupload"
  | "clouddownload"
  | "clouddownloado"
  | "clouduploado"
  | "enviroment"
  | "enviromento"
  | "eye"
  | "eyeo"
  | "camera"
  | "camerao"
  | "windows"
  | "export2"
  | "export"
  | "circledowno"
  | "circledown"
  | "hdd"
  | "ie"
  | "delete"
  | "enter"
  | "pushpino"
  | "pushpin"
  | "heart"
  | "hearto"
  | "smile-circle"
  | "smileo"
  | "frowno"
  | "calculator"
  | "chrome"
  | "github"
  | "iconfontdesktop"
  | "caretcircleoup"
  | "upload"
  | "download"
  | "piechart"
  | "lock1"
  | "unlock"
  | "windowso"
  | "dotchart"
  | "barchart"
  | "codesquare"
  | "plussquare"
  | "minussquare"
  | "closesquare"
  | "closesquareo"
  | "checksquare"
  | "checksquareo"
  | "fastbackward"
  | "fastforward"
  | "upsquare"
  | "downsquare"
  | "leftsquare"
  | "rightsquare"
  | "rightsquareo"
  | "leftsquareo"
  | "down-square-o"
  | "up-square-o"
  | "play"
  | "playcircleo"
  | "tag"
  | "tago"
  | "addfile"
  | "folder1"
  | "file1"
  | "switcher"
  | "addfolder"
  | "folderopen"
  | "search1"
  | "ellipsis1"
  | "calendar"
  | "filetext1"
  | "copy1"
  | "jpgfile1"
  | "pdffile1"
  | "exclefile1"
  | "pptfile1"
  | "unknowfile1"
  | "wordfile1"
  | "dingding"
  | "dingding-o"
  | "mobile1"
  | "tablet1"
  | "bells"
  | "disconnect"
  | "database"
  | "barcode"
  | "hourglass"
  | "key"
  | "flag"
  | "layout"
  | "printer"
  | "USB"
  | "skin"
  | "tool"
  | "car"
  | "addusergroup"
  | "carryout"
  | "deleteuser"
  | "deleteusergroup"
  | "man"
  | "isv"
  | "gift"
  | "idcard"
  | "medicinebox"
  | "redenvelopes"
  | "rest"
  | "Safety"
  | "wallet"
  | "woman"
  | "adduser"
  | "bank"
  | "Trophy"
  | "loading1"
  | "loading2"
  | "like2"
  | "dislike2"
  | "like1"
  | "dislike1"
  | "bulb1"
  | "rocket1"
  | "select1"
  | "apple1"
  | "apple-o"
  | "android1"
  | "android"
  | "aliwangwang-o1"
  | "aliwangwang"
  | "pay-circle1"
  | "pay-circle-o1"
  | "poweroff"
  | "trademark"
  | "find"
  | "copyright"
  | "sound"
  | "earth"
  | "wifi"
  | "sync"
  | "login"
  | "logout"
  | "reload1"
  | "message1"
  | "shake"
  | "API"
  | "appstore-o"
  | "appstore1"
  | "scan1"
  | "exception1"
  | "contacts"
  | "solution1"
  | "fork"
  | "edit"
  | "form"
  | "warning"
  | "table"
  | "profile"
  | "dashboard"
  | "indent-left"
  | "indent-right"
  | "menu-unfold"
  | "menu-fold"
  | "antdesign"
  | "alipay-square"
  | "codepen-circle"
  | "google"
  | "amazon"
  | "codepen"
  | "facebook-square"
  | "dropbox"
  | "googleplus"
  | "linkedin-square"
  | "medium-monogram"
  | "gitlab"
  | "medium-wordmark"
  | "QQ"
  | "skype"
  | "taobao-square"
  | "alipay-circle"
  | "youtube"
  | "wechat"
  | "twitter"
  | "weibo"
  | "HTML"
  | "taobao-circle"
  | "weibo-circle"
  | "weibo-square"
  | "CodeSandbox"
  | "aliyun"
  | "zhihu"
  | "behance"
  | "dribbble"
  | "dribbble-square"
  | "behance-square"
  | "file-markdown"
  | "instagram"
  | "yuque"
  | "slack"
  | "slack-square";
// ...
export type EvilIconsNames =
  | "archive"
  | "arrow-down"
  | "arrow-left"
  | "arrow-right"
  | "arrow-up"
  | "bell"
  | "calendar"
  | "camera"
  | "cart"
  | "chart"
  | "check"
  | "chevron-down"
  | "chevron-left"
  | "chevron-right"
  | "chevron-up"
  | "clock"
  | "close"
  | "close-o"
  | "comment"
  | "credit-card"
  | "envelope"
  | "exclamation"
  | "external-link"
  | "eye"
  | "gear"
  | "heart"
  | "image"
  | "like"
  | "link"
  | "location"
  | "lock"
  | "minus"
  | "navicon"
  | "paperclip"
  | "pencil"
  | "play"
  | "plus"
  | "pointer"
  | "question"
  | "redo"
  | "refresh"
  | "retweet"
  | "sc-facebook"
  | "sc-github"
  | "sc-google-plus"
  | "sc-instagram"
  | "sc-linkedin"
  | "sc-odnoklassniki"
  | "sc-pinterest"
  | "sc-skype"
  | "sc-soundcloud"
  | "sc-telegram"
  | "sc-tumblr"
  | "sc-twitter"
  | "sc-vimeo"
  | "sc-vk"
  | "sc-youtube"
  | "search"
  | "share-apple"
  | "share-google"
  | "spinner"
  | "spinner-2"
  | "spinner-3"
  | "star"
  | "tag"
  | "trash"
  | "trophy"
  | "undo"
  | "unlock"
  | "user";

export type FeatherNames =
  | "activity"
  | "airplay"
  | "alert-circle"
  | "alert-octagon"
  | "alert-triangle"
  | "align-center"
  | "align-justify"
  | "align-left"
  | "align-right"
  | "anchor"
  | "aperture"
  | "archive"
  | "arrow-down"
  | "arrow-down-circle"
  | "arrow-down-left"
  | "arrow-down-right"
  | "arrow-left"
  | "arrow-left-circle"
  | "arrow-right"
  | "arrow-right-circle"
  | "arrow-up"
  | "arrow-up-circle"
  | "arrow-up-left"
  | "arrow-up-right"
  | "at-sign"
  | "award"
  | "bar-chart"
  | "bar-chart-2"
  | "battery"
  | "battery-charging"
  | "bell"
  | "bell-off"
  | "bluetooth"
  | "bold"
  | "book"
  | "book-open"
  | "bookmark"
  | "box"
  | "briefcase"
  | "calendar"
  | "camera"
  | "camera-off"
  | "cast"
  | "check"
  | "check-circle"
  | "check-square"
  | "chevron-down"
  | "chevron-left"
  | "chevron-right"
  | "chevron-up"
  | "chevrons-down"
  | "chevrons-left"
  | "chevrons-right"
  | "chevrons-up"
  | "chrome"
  | "circle"
  | "clipboard"
  | "clock"
  | "cloud"
  | "cloud-drizzle"
  | "cloud-lightning"
  | "cloud-off"
  | "cloud-rain"
  | "cloud-snow"
  | "code"
  | "codepen"
  | "codesandbox"
  | "coffee"
  | "columns"
  | "command"
  | "compass"
  | "copy"
  | "corner-down-left"
  | "corner-down-right"
  | "corner-left-down"
  | "corner-left-up"
  | "corner-right-down"
  | "corner-right-up"
  | "corner-up-left"
  | "corner-up-right"
  | "cpu"
  | "credit-card"
  | "crop"
  | "crosshair"
  | "database"
  | "delete"
  | "disc"
  | "divide"
  | "divide-circle"
  | "divide-square"
  | "dollar-sign"
  | "download"
  | "download-cloud"
  | "dribbble"
  | "droplet"
  | "edit"
  | "edit-2"
  | "edit-3"
  | "external-link"
  | "eye"
  | "eye-off"
  | "facebook"
  | "fast-forward"
  | "feather"
  | "figma"
  | "file"
  | "file-minus"
  | "file-plus"
  | "file-text"
  | "film"
  | "filter"
  | "flag"
  | "folder"
  | "folder-minus"
  | "folder-plus"
  | "framer"
  | "frown"
  | "gift"
  | "git-branch"
  | "git-commit"
  | "git-merge"
  | "git-pull-request"
  | "github"
  | "gitlab"
  | "globe"
  | "grid"
  | "hard-drive"
  | "hash"
  | "headphones"
  | "heart"
  | "help-circle"
  | "hexagon"
  | "home"
  | "image"
  | "inbox"
  | "info"
  | "instagram"
  | "italic"
  | "key"
  | "layers"
  | "layout"
  | "life-buoy"
  | "link"
  | "link-2"
  | "linkedin"
  | "list"
  | "loader"
  | "lock"
  | "log-in"
  | "log-out"
  | "mail"
  | "map"
  | "map-pin"
  | "maximize"
  | "maximize-2"
  | "meh"
  | "menu"
  | "message-circle"
  | "message-square"
  | "mic"
  | "mic-off"
  | "minimize"
  | "minimize-2"
  | "minus"
  | "minus-circle"
  | "minus-square"
  | "monitor"
  | "moon"
  | "more-horizontal"
  | "more-vertical"
  | "mouse-pointer"
  | "move"
  | "music"
  | "navigation"
  | "navigation-2"
  | "octagon"
  | "package"
  | "paperclip"
  | "pause"
  | "pause-circle"
  | "pen-tool"
  | "percent"
  | "phone"
  | "phone-call"
  | "phone-forwarded"
  | "phone-incoming"
  | "phone-missed"
  | "phone-off"
  | "phone-outgoing"
  | "pie-chart"
  | "play"
  | "play-circle"
  | "plus"
  | "plus-circle"
  | "plus-square"
  | "pocket"
  | "power"
  | "printer"
  | "radio"
  | "refresh-ccw"
  | "refresh-cw"
  | "repeat"
  | "rewind"
  | "rotate-ccw"
  | "rotate-cw"
  | "rss"
  | "save"
  | "scissors"
  | "search"
  | "send"
  | "server"
  | "settings"
  | "share"
  | "share-2"
  | "shield"
  | "shield-off"
  | "shopping-bag"
  | "shopping-cart"
  | "shuffle"
  | "sidebar"
  | "skip-back"
  | "skip-forward"
  | "slack"
  | "slash"
  | "sliders"
  | "smartphone"
  | "smile"
  | "speaker"
  | "square"
  | "star"
  | "stop-circle"
  | "sun"
  | "sunrise"
  | "sunset"
  | "tablet"
  | "tag"
  | "target"
  | "terminal"
  | "thermometer"
  | "thumbs-down"
  | "thumbs-up"
  | "toggle-left"
  | "toggle-right"
  | "tool"
  | "trash"
  | "trash-2"
  | "trello"
  | "trending-down"
  | "trending-up"
  | "triangle"
  | "truck"
  | "tv"
  | "twitch"
  | "twitter"
  | "type"
  | "umbrella"
  | "underline"
  | "unlock"
  | "upload"
  | "upload-cloud"
  | "user"
  | "user-check"
  | "user-minus"
  | "user-plus"
  | "user-x"
  | "users"
  | "video"
  | "video-off"
  | "voicemail"
  | "volume"
  | "volume-1"
  | "volume-2"
  | "volume-x"
  | "watch"
  | "wifi"
  | "wifi-off"
  | "wind"
  | "x"
  | "x-circle"
  | "x-octagon"
  | "x-square"
  | "youtube"
  | "zap"
  | "zap-off"
  | "zoom-in"
  | "zoom-out";
// ...
export type IconLibrary =
  | "AntDesign"
  | "Entypo"
  | "EvilIcons"
  | "Feather"
  | "FontAwesome"
  | "FontAwesome5"
  | "FontAwesome5Pro"
  | "FontAwesome6"
  | "FontAwesome6Pro"
  | "Fontisto"
  | "Foundation"
  | "Ionicons"
  | "MaterialCommunityIcons"
  | "MaterialIcons"
  | "Octicons"
  | "SimpleLineIcons"
  | "Zocial";

export type IconNames<T extends IconLibrary> = T extends "AntDesign"
  ? AntDesignNames
  : T extends "Entypo"
  ? EntypoNames
  : T extends "EvilIcons"
  ? EvilIconsNames
  : T extends "Feather"
  ? FeatherNames
  : T extends "FontAwesome"
  ? FontAwesomeNames
  : T extends "FontAwesome5"
  ? FontAwesome5Names
  : T extends "FontAwesome5BrandsNames"
  ? FontAwesome5BrandsNames
  : T extends "FontAwesome6"
  ? FontAwesome6Names
  : T extends "FontAwesome6Pro"
  ? FontAwesome6BrandsNames
  : T extends "Fontisto"
  ? FontistoNames
  : T extends "Foundation"
  ? FoundationNames
  : T extends "Ionicons"
  ? IoniconsNames
  : T extends "MaterialCommunityIcons"
  ? MaterialCommunityIconsNames
  : T extends "MaterialIcons"
  ? MaterialIconsNames
  : T extends "Octicons"
  ? OcticonsNames
  : T extends "SimpleLineIcons"
  ? SimpleLineIconsNames
  : T extends "Zocial"
  ? ZocialNames
  : never;